Rename pv_slicer -> path_slicer; thin wrappers around glue-core
